// src/types.d.ts
export type userResponse = {
    users: User[],
    numItmes: number
}

export type User = {
    id: string,
    numIdent: string,
    firstName: string,
    lastName: string,
    userName: string,
    image: string,
    email: string,
    phone: string,
    status: boolean,
    clientId: any,
    roleId: string,
    identId: string,
    Identification: {
        name: string
    }
    Role: {
        role: string
    }
    createdAt:string
}

export type Session = {
    msg: string
    token: string
    user: User
    success: boolean
}

export type Roles = {
    roles: Role[],
    success: boolean
}

export type Role = {
    role: string,
    id: string
}
export type IdentificationResponse = {
    identifications: IndentificationType[],
    success: boolean
}

export type IndentificationType = {
    name: string,
    id: string
}